requirements-sim.txt
google-cloud-storage==3.3.1
python_dateutil==2.9.0.post0
python-dotenv==1.2.2
pyarrow